+++ src/models/operations/index.ts
@@ -12,4 +12,3 @@ export * from "./getpetbyids.js";
 export * from "./getuserbyname.js";
 export * from "./saynewrandom.js";
 export * from "./updateuser.js";
-export * from "./uploadfile.js";