+++ RELEASES.md
@@ -683,3 +683,13 @@ Based on:
 - [typescript v5.2.0-alpha.40] .
 ### Releases
 - [NPM v5.2.0-alpha.40] https://www.npmjs.com/package/meep-moop/v/5.2.0-alpha.40 - .
+
+## 2025-08-08 12:09:57
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.599.0 (2.675.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v5.2.0-alpha.44] .
+### Releases
+- [NPM v5.2.0-alpha.44] https://www.npmjs.com/package/meep-moop/v/5.2.0-alpha.44 - .